feat: unite hooks related to genres into one hook and define update hook
